back/auth-service/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
asyncpg>=0.25.0
email-validator>=1.1.0
alembic>=1.12.0
psycopg2-binary